requirements.txt

viam-sdk==0.71.0

typing-extensions

aiohttp
